4 structure the api response as per best practices

+++ SportsZone/SportsZoneWebAPI/Controllers/CustomerController.cs
@@ -27,7 +27,7 @@ namespace SportsZoneWebAPI.Controllers
             }
             catch (Exception e)
             {
-                return BadRequest(e.Message);
+                return StatusCode(500, e.Message);
             }
         }
         [HttpGet, Route("GetCustomerByCustomerID/{email}")]
@@ -35,12 +35,21 @@ namespace SportsZoneWebAPI.Controllers
         {
             try
             {
+                if (string.IsNullOrWhiteSpace(email))
+                {
+                    // Invalid input: Return a BadRequest response with an error message
+                    return BadRequest();
+                }
+                if (!await _customerService.IsAvail(email))
+                {
+                    return NotFound();
+                }
                 CustomerResponseDTO customerResponseDTO = await _customerService.GetCustomerByCustomerID(email);
                 return Ok(customerResponseDTO);
             }
             catch (Exception e)
             {
-                return BadRequest(e.Message);
+                return StatusCode(500, e.Message);
             }
         }
 
@@ -49,12 +58,21 @@ namespace SportsZoneWebAPI.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                {
+                    // Model validation failed: Return a BadRequest response with validation errors
+                    return BadRequest(ModelState);
+                }
+                if (await _customerService.IsAvail(customerRequestDTO.Email))
+                {
+                    return Conflict();
+                }
                 await _customerService.CreateCustomer(customerRequestDTO);
                 return Ok(customerRequestDTO);
             }
             catch (Exception e)
             {
-                return BadRequest(e.Message);
+                return StatusCode(500, e.Message);
             }
         }
 
@@ -63,41 +81,62 @@ namespace SportsZoneWebAPI.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                {
+                    // Model validation failed: Return a BadRequest response with validation errors
+                    return BadRequest(ModelState);
+                }
+                if (!await _customerService.IsAvail(customerRequestDTO.Email))
+                {
+                    return NotFound();
+                }
                 await _customerService.UpdateCustomer(customerRequestDTO);
                 return Ok(customerRequestDTO);
             }
             catch (Exception e)
             {
-                return BadRequest(e.Message);
+                return StatusCode(500, e.Message);
             }
         }
-        [HttpDelete, Route("DeleteAllCustomers")]
-        public async Task<ActionResult> DeleteAllCustomers()
-        {
-            try
-            {
-                await _customerService.DeleteAllCustomers();
-                return Ok("All Customers deleted succesfully");
-            }
-            catch (Exception e)
-            {
-                return BadRequest(e.Message);
-            }
-        }
+
 
         [HttpDelete, Route("DeleteCustomerByCustomerID/{email}")]
         public async Task<ActionResult> DeleteCustomerByCustomerID(string email)
         {
             try
             {
+                if (string.IsNullOrWhiteSpace(email))
+                {
+                    // Invalid input: Return a BadRequest response with an error message
+                    return BadRequest("Input parameter 'email' is required and cannot be empty.");
+                }
+                if (!await _customerService.IsAvail(email))
+                {
+                    return NotFound();
+                }
                 await _customerService.DeleteCustomerByCustomerID(email);
                 return Ok($"Customer with ID : {email} deleted succesfully");
             }
             catch (Exception e)
             {
-                return BadRequest(e.Message);
+                return StatusCode(500, e.Message);
             }
         }
 
+        #region Rare use case
+        //[HttpDelete, Route("DeleteAllCustomers")]
+        //public async Task<ActionResult> DeleteAllCustomers()
+        //{
+        //    try
+        //    {
+        //        await _customerService.DeleteAllCustomers();
+        //        return Ok("All Customers deleted succesfully");
+        //    }
+        //    catch (Exception e)
+        //    {
+        //        return StatusCode(500, e.Message);
+        //    }
+        //} 
+        #endregion
     }
 }
